AtomNetwork Reconfig
